Rewrite external link svg to another colour
As this is rendered not within the page but rather on its own we don't
have any css var information etc. to set the fill based on current
styling. A changing svg could be achieved via @media in svg style but
this breaks the manual switching.
Settling now for #666666 which is distinguishable both in dark an light
just enough to satisfy my view. Probably horrible by any means.