Rearrange and rewrite include project-info
Languages and tools now come before the generic info since they are more
like the category listing at the beginning. Generic info now starts with
the status of the project to act like a label to the menu. Every other
entry in generic is now prepared to be styled like a hollow button to
signal that they are clickable.